refactor: 리뷰 리포지토리 추상화
