+++ src/modules/base/essay/core/essay.service.ts
@@ -8,7 +8,7 @@ import { Story } from '../../../../entities/story.entity';
 import { User } from '../../../../entities/user.entity';
 import { ToolService } from '../../../utils/tool/tool.service';
 import { AwsService } from '../../../adapters/aws/core/aws.service';
-import { ReviewService } from '../../../extensions/management/review/review.service';
+import { ReviewService } from '../../../extensions/management/review/core/review.service';
 import { StoryService } from '../../../extensions/essay/story/core/story.service';
 import { UserService } from '../../user/core/user.service';
 import { TagService } from '../../../extensions/essay/tag/core/tag.service';
