+++ src/modules/extensions/management/report/infrastructure/report.repository.ts
@@ -1,8 +1,9 @@
 import { Repository } from 'typeorm';
-import { ReportQueue } from '../../../../entities/reportQueue.entity';
+import { ReportQueue } from '../../../../../entities/reportQueue.entity';
 import { InjectRepository } from '@nestjs/typeorm';
+import { IReporter } from './ireport.repository';
 
-export class ReportRepository {
+export class ReportRepository implements IReporter {
   constructor(
     @InjectRepository(ReportQueue)
     private readonly reportRepository: Repository<ReportQueue>,
