+++ src/modules/extensions/management/report/report.module.ts
@@ -3,10 +3,10 @@ import { forwardRef, Module } from '@nestjs/common';
 import { User } from '../../../../entities/user.entity';
 import { Essay } from '../../../../entities/essay.entity';
 import { ReportQueue } from '../../../../entities/reportQueue.entity';
-import { ReportService } from './report.service';
-import { ReportRepository } from './report.repository';
+import { ReportService } from './core/report.service';
+import { ReportRepository } from './infrastructure/report.repository';
 import { EssayModule } from '../../../base/essay/essay.module';
-import { ReportController } from './report.controller';
+import { ReportController } from './api/report.controller';
 import { JwtModule } from '@nestjs/jwt';
 import { AuthModule } from '../../../base/auth/auth.module';
 import { UserModule } from '../../../base/user/user.module';
@@ -20,7 +20,7 @@ import { UserModule } from '../../../base/user/user.module';
     forwardRef(() => EssayModule),
   ],
   controllers: [ReportController],
-  providers: [ReportService, ReportRepository],
-  exports: [ReportService, ReportRepository],
+  providers: [ReportService, { provide: 'IReporter', useClass: ReportRepository }],
+  exports: [ReportService, { provide: 'IReporter', useClass: ReportRepository }],
 })
 export class ReportModule {}
